pgduck_server: move the inode logic into its own file
file_cache_manager.cpp was already long, and the inode side of cache
management is a self-contained concern: the setting, what the file system
reports, the floor we derive from it, and the empty-directory pruning that
gives inodes back. Move all of that to fs/cache_inode_budget.{cpp,hpp},
along with the InodeBudget struct that ManageCache carries around.

No behaviour change. What stays in file_cache_manager.cpp is the policy
that combines bytes and inodes: the eviction loop, SkipQueuedDownloads,
and the cache action bookkeeping. CheckMinFreeInodes was a static member
of FileCacheManager only to keep it near the setting, so it becomes a free
CheckMinFreeCacheInodes, and the constants that nothing outside the
implementation used are now file-internal.
